refactor(lib): split resolve-setup.ts into four more focused modules
Continuing the decomposition started in #116. Pull four more
self-contained sections out of resolve-setup.ts:

* `src/lib/raw-inputs.ts` — `readRawInputs` (~37 LOC). Pure env →
  RawInputs mapping; each entry is one row of action.yml inputs.
* `src/lib/input-parsers.ts` — small pure parsers and normalizers:
  `parseCacheShutdownOnIdleSeconds`, `detectUserLinkerEnv`,
  `normalizeStatsMode`, `normalizeCompileCacheStats` (~60 LOC).
* `src/lib/fetch-release.ts` — `fetchReleaseTagDefault`,
  `resolveSoldrReleaseVersion`, and the `ResolveSetupDeps` interface.
  Isolates the GitHub REST + auth wiring (~75 LOC).
* `src/lib/python-json.ts` — `pythonDefaultJson` and its private
  formatter. Mirrors Python's `json.dumps(..., sort_keys=True)` for
  the toolchain-signature digest (~30 LOC).

`resolve-setup.ts` re-exports every public symbol so existing test
and main.ts imports keep working without a single edit. The remaining
~700 lines in `resolve-setup.ts` are now the actual orchestration:
cache-key derivation, plan assembly, and the resolveSetup +
applyResolveResult exports.

`resolve-setup.ts`: 898 → 707 lines.

Pure refactor — no behavior change. Full suite stays at 201 pass /
4 skipped.
